test(meta): group view event tests
