+++ src/common/meta/src/ddl/tests/event.rs
@@ -14,3 +14,4 @@
 
 mod database;
 mod flow;
+mod view;
